backend/pyproject.toml
[project]
name = "link-shortener-server"
version = "0.1.0"
description = "Link Shortener backend server"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "redis[hiredis]>=5.3.0",
    "redis-om>=0.3.5",
]
